added comments to reaction db model and api route

+++ api/routes/reaction.py
@@ -8,7 +8,13 @@ from api.database.models import Reaction
 
 @app.route("/reaction", methods=["GET", "POST", "PATCH", "DELETE"])
 def reaction():
+
     if request.method == "POST":
+        """
+        Reaction [POST] request - adds a new reaction entry into the db
+        Input:  Requires the account name, course name, and rating (optional) as a form
+        Output: Returns the entry that was created.
+        """
         account = request.form["account"]
         course = request.form["course"]
         rating = request.form["rating"]
@@ -26,6 +32,11 @@ def reaction():
         return jsonify(new_entry.serialize())
 
     elif request.method == "GET":
+        """
+        Reaction [GET] request - returns the views and average rating of a course
+        Input:  Requires the course name as an argument/parameter
+        Output: Returns the number of views for the specific course and the average rating
+        """
         course = request.args.get("course")
         result = Reaction.query.filter_by(
             course_name=course,
@@ -42,6 +53,11 @@ def reaction():
         return json.dumps([views, finalrating])
 
     elif request.method == "PATCH":
+        """
+        Reaction [PATCH] request - updates an existing entry with a new rating
+        Input:  Requires the account name, course name as an argument/parameter and the new rating as a form
+        Output: Returns the entry with the updated rating - should only be 1 item
+        """
         account = request.args.get("account")
         course = request.args.get("course")
         rating = request.form["rating"]
@@ -53,19 +69,18 @@ def reaction():
         return jsonify([item.serialize() for item in entry])
 
     elif request.method == "DELETE":
+        """
+        Reaction [DELETE] request - deletes an entry from the db
+        Input:  Requires the account name, course name as an argument/parameter
+        Output: Returns the number of items deleted
+        """
         account = request.args.get("account")
         course = request.args.get("course")
 
-        if account is not None:
-            if course is None:
-                result = Reaction.query.filter_by(
-                    account_name=account,
-                ).delete()
-            else:
-                result = Reaction.query.filter_by(
-                    account_name=account,
-                    course_name=course,
-                ).delete()
+        if account is not None and course is None:
+            result = Reaction.query.filter_by(
+                account_name=account,
+            ).delete()
 
         db.session.commit()
         return jsonify(result)
